Make it easier to integrate dailydrip exports
So this is just a hack to reduce the time it takes me to update the 2
sites, but it's ultimately going to be ok.  I'm building some tools to
produce the static files middleman needs too.

WIP commit because if I have to upgrade middleman bad stuff will go down
re: dependencies sigh